+++ src/utils/types_utils_test.ts
@@ -8,10 +8,6 @@
  * =============================================================================
  */
 
-import {scalar, zeros} from '@tensorflow/tfjs-core';
-
-import {LayerVariable} from '../variables';
-
 import * as types_utils from './types_utils';
 
 describe('isArrayOfShapes', () => {
@@ -63,35 +59,5 @@ describe('getExactlyOneShape', () => {
     expect(() => types_utils.getExactlyOneShape([
       [1], [2]
     ])).toThrowError(/Expected exactly 1 Shape; got 2/);
-  });
-});
-
-describe('countParamsInWeights', () => {
-  it('Zero weights', () => {
-    expect(types_utils.countParamsInWeights([])).toEqual(0);
-  });
-
-  it('One float32 weight', () => {
-    const weight1 = new LayerVariable(zeros([2, 3]));
-    expect(types_utils.countParamsInWeights([weight1])).toEqual(6);
-  });
-
-  it('One float32 scalar weight', () => {
-    const weight1 = new LayerVariable(scalar(42));
-    expect(types_utils.countParamsInWeights([weight1])).toEqual(1);
-  });
-
-  it('One int32 weight', () => {
-    const weight1 = new LayerVariable(zeros([1, 3, 4], 'int32'), 'int32');
-    expect(types_utils.countParamsInWeights([weight1])).toEqual(12);
-  });
-
-  it('Two weights, mixed types and shapes', () => {
-    const weight1 = new LayerVariable(scalar(42));
-    const weight2 = new LayerVariable(zeros([2, 3]));
-    const weight3 = new LayerVariable(zeros([1, 3, 4], 'int32'), 'int32');
-    expect(types_utils.countParamsInWeights([
-      weight1, weight2, weight3
-    ])).toEqual(19);
   });
 });
